Rename file from kebab case to upper camel
